refactor: 557 - move numbers to separate ui folder
