+++ src/widgets/numbers/ui/numbers/numbers.module.scss
