+++ src/widgets/numbers/ui/numbers/numbers.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { beforeEach, describe, expect, it } from 'vitest';
-import { Numbers } from './ui/numbers';
+import { Numbers } from '../numbers/numbers.tsx';
 
 describe('Numbers', () => {
   beforeEach(() => {
